requirements.txt
amqp==5.3.1
asgiref==3.11.0
asttokens==3.0.1
async-timeout==5.0.1
billiard==4.2.3
celery==5.5.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
decorator==5.2.1
Django==5.2.8
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
executing==2.2.1
idna==3.11
ipython_pygments_lexers==1.1.1
jedi==0.19.2
kombu==5.5.4
matplotlib-inline==0.2.1
mssql-django==1.6
packaging==25.0
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
PyJWT==2.11.0
pyodbc==5.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==7.1.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
